+++ R/get_concept.R
@@ -1,45 +1,45 @@
 #' Get a concept in an ontology
 #'
-#' @param x [`character(1)`][character]\cr a table containing all columns of the
+#' @param table [`character(1)`][character]\cr a table containing all columns (a
+#'   subset of "id", "class", "label", "has_broader" and "has_source") of the
 #'   ontology that shall be filter by the values in those columns.
-#' @param ... combination of column name and value to filter that column by. The
-#'   value to filter by can be provided as regular expression.
-#' @param regex [`logical(1)`][logical]\cr whether or not the value in
-#'   \code{...} shall be matched in full, or whether any partial match should be
-#'   returned.
-#' @param tree [`logical(1)`][logical]\cr whether or not to output the whole
-#'   ontology tree starting from the given search terms.
-#' @param na.rm [`logical(1)`][logical]\cr whether or not missing concepts are
-#'   omitted or not.
-#' @param mappings [`character(1)`][character]\cr the type of mappings within
-#'   which to search for the concepts, possible values are \code{"none"},
-#'   \code{"all"}, \code{"close"}, \code{"broader"}, \code{"narrower"},
-#'   \code{"exact"} or any combination thereof.
 #' @param ontology [`ontology(1)`][list]\cr either a path where the ontology is
 #'   stored, or an already loaded ontology.
+#' @param per_class [`logical(1)`][logical]\cr whether ot not to flatten the
+#'   ontology before matching \code{table} with the ontology, whereby
+#'   \code{table} would contain columns of the classes in the ontology. This can
+#'   be useful when concepts are unique only within their parent concepts, so
+#'   that unique identification is only possible when they are matched together.
+#' @param regex [`logical(1)`][logical]\cr if \code{regex = TRUE}, the columns
+#'   defined in \code{table} are filtered by \code{\link[stringr]{str_detect}}
+#'   on the column values (if you define several, they are combined with an AND
+#'   operator), otherwise a \code{\link[dplyr]{left_join}} on the ontology is
+#'   carried out.
+#' @param mappings [`logical(1)`][logical]\cr whether or not to provide a table
+#'   that includes mappings. In this case, only unique items of the concepts in
+#'   \code{table} are included in the output table.
+#' @param external [`logical(1)`][logical]\cr whether or not to return merely
+#'   the table of external concepts.
 #' @examples
 #' ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
 #' onto <- load_ontology(path = ontoDir)
 #'
 #' # exact matches from a loaded ontology ...
-#' get_concept(x = data.frame(label = "FODDER CROPS"), ontology = onto)
+#' get_concept(table = data.frame(label = "FODDER CROPS"), ontology = onto)
 #'
-#' # ... or one stored on the harddisc
-#' get_concept(x = data.frame(label = "FODDER CROPS"), ontology = ontoDir)
+#' # ... or one stored on the harddisc, with mappings (none available here)
+#' get_concept(table = data.frame(label = "FODDER CROPS"), ontology = ontoDir,
+#'             mappings = TRUE)
 #'
-#' # use regular expressions ...
-#' get_concept(x = data.frame(label = "/*crops"), regex = TRUE, ontology = onto)
-#'
-#' get_concept(x = data.frame(label = "/*crops"), has_broader = ".05", regex = TRUE, ontology = onto)
-#'
-#' # get all concepts that are nested into another concept
-#' get_concept(x = data.frame(label = "FODDER CROPS"), tree = TRUE, ontology = onto)
+#' # extract concepts based on regular expressions
+#' get_concept(table = data.frame(label = "crop", id = ".03$"),
+#'             regex = TRUE, ontology = ontoDir)
 #' @return A table of a subset of the ontology according to the values in
 #'   \code{...}
 #' @importFrom checkmate assertFileExists assertLogical testChoice
 #' @importFrom tibble as_tibble
 #' @importFrom readr read_rds
-#' @importFrom tidyselect everything
+#' @importFrom tidyselect everything contains
 #' @importFrom tidyr separate_rows separate pivot_longer pivot_wider
 #' @importFrom rlang quos eval_tidy := sym as_name
 #' @importFrom dplyr filter pull select rename inner_join
@@ -49,23 +49,17 @@
 #' @importFrom utils head
 #' @export
 
-get_concept <- function(x = NULL, ..., regex = FALSE, tree = FALSE,
-                        na.rm = FALSE, mappings = "none", ontology = NULL){
+get_concept <- function(table = NULL, ontology = NULL, per_class = FALSE,
+                        mappings = FALSE, regex = FALSE, external = FALSE
+                        ){
 
-  assertDataFrame(x = x, null.ok = FALSE)
+  # mappings = FALSE; regex = FALSE; external = FALSE
+
+  assertDataFrame(x = table, null.ok = TRUE)
+  assertLogical(x = mappings, len = 1, any.missing = FALSE)
   assertLogical(x = regex, len = 1, any.missing = FALSE)
-  assertLogical(x = tree, len = 1, any.missing = FALSE)
-  assertLogical(x = na.rm, len = 1, any.missing = FALSE)
-  assertSubset(x = mappings, choices = c("all", "none", "close", "broader", "narrower", "exact"))
-
-  if("none" %in% mappings){
-    mappings <- NULL
-  } else {
-    if("all" %in% mappings){
-      mappings <- c("close", "broader", "narrower", "exact")
-    }
-    mappings <- paste0("has_", mappings, "_match")
-  }
+  assertLogical(x = external, len = 1, any.missing = FALSE)
+  assertLogical(x = per_class, len = 1, any.missing = FALSE)
 
   if(!inherits(x = ontology, what = "onto")){
     assertFileExists(x = ontology, access = "r", extension = "rds")
@@ -76,134 +70,207 @@ get_concept <- function(x = NULL, ..., regex = FALSE, tree = FALSE,
     ontology <- load_ontology(path = ontoPath)
   }
 
-  theConcepts <- ontology@concepts
 
-  attrib <- quos(..., .named = TRUE)
-  # return(attrib)
+  if(external){
 
-  # identify attributes that are not in the ontology
-  if(!all(names(attrib) %in% colnames(theConcepts$harmonised))){
-    sbst <- names(attrib) %in% colnames(theConcepts$harmonised)
-    theName <- names(attrib)[!sbst]
-    warning(paste0("'", paste0(theName, collapse = ", "), "' is not a column in the ontology and is thus ignored."))
-    attrib <- attrib[sbst]
-  }
-
-  if(regex){
-
-    if(!is.null(x)){
-      toOut <- ontology@concepts$harmonised %>%
-        filter(str_detect(label, x$label))
+    if(is.null(table)){
+      toOut <- ontology@concepts$external
     } else {
-      toOut <- ontology@concepts$harmonised
+
+      assertNames(x = names(table), subset.of = c("id", "label", "description", "has_broader", "has_source"))
+
+      if(!regex){
+
+        toOut <- table %>%
+          left_join(ontology@concepts$external, by = colnames(table)) %>%
+          mutate(external = label,
+                 match = "exact",
+                 has_source = "1") %>%
+          select(external, match, label, id, has_broader, description, has_source)
+
+      } else {
+
+        toOut <- ontology@concepts$external
+
+        for(j in seq_along(table)){
+          toOut <- toOut %>%
+            filter(str_detect(string = toOut[[colnames(table)[j]]],
+                              pattern = paste0(c(table %>% pull(j)), collapse = "|")))
+        }
+
+      }
     }
 
-    for(i in seq_along(attrib)){
-
-      toOut <- toOut %>%
-        filter(str_detect(toOut[[names(attrib)[i]]], paste0(as_name(attrib[[i]]), collapse = "|")))
-
-    }
 
   } else {
 
-    assertNames(x = names(x), subset.of = c("id", "has_broader", "source_id", "class", "label", "source_label", "external_label"))
+    if(is.null(table)){
+      toOut <- ontology@concepts$harmonised
+    } else {
 
-    # get the already harmonised concepts
-    toOut <- x %>%
-      left_join(theConcepts$harmonised, by = colnames(x))
+      if(per_class){
 
-    if(!is.null(mappings)){
+        theClasses <- unique(ontology@concepts$harmonised$class)
+        assertNames(x = names(table), subset.of = theClasses)
 
-      # and replace external IDs with the labels
-      toOut <- toOut%>%
-        pivot_longer(cols = c(has_close_match, has_broader_match, has_narrower_match, has_exact_match), names_to = "match", values_to = "extid") %>%
-        separate_rows(extid, sep = " \\| ") %>%
-        separate(col = extid, into = c("extid", "certainty"), sep = "[.]") %>%
-        left_join(theConcepts$external %>% select(extid = id, external_label = label), by = "extid") %>%
-        group_by(label, class, id, description, has_broader, match) %>%
-        summarise(external_label = paste0(external_label, collapse = " | ")) %>%
-        ungroup() %>%
-        mutate(external_label = na_if(external_label, "NA")) %>%
-        pivot_wider(id_cols = c(label, class, id, description, has_broader), names_from = match, values_from = external_label)
+        flatOnto <- ontology@concepts$harmonised %>%
+          filter(class == theClasses[1])
 
-      toOutCompl <- toOut %>%
-        filter(!is.na(id))
+        for(i in seq_along(theClasses)){
+          if(i == length(theClasses)) break
 
-      matchCols <- colnames(x)[colnames(x) %in% colnames(theConcepts$external)]
-      toMatch <- x %>%
-        left_join(theConcepts$external, matchCols) %>%
-        filter(!is.na(id)) %>%
-        filter(!label %in% toOutCompl$label)
+          temp <- ontology@concepts$harmonised %>%
+            filter(class == theClasses[i]) %>%
+            select(id, !!theClasses[i] := label)
 
-      if(dim(toMatch)[1] != 0){
-        temp <- theConcepts$harmonised %>%
-          pivot_longer(cols = mappings, names_to = "match", values_to = "extid") %>%
-          filter(!is.na(extid)) %>%
-          separate_rows(extid, sep = " \\| ") %>%
-          separate(col = extid, into = c("extid", "certainty"), sep = "[.]") %>%
-          left_join(toMatch %>% select(extid = id, external_label = label), by = "extid") %>%
-          filter(!is.na(external_label))
+          child <- ontology@concepts$harmonised %>%
+            filter(class == theClasses[i + 1]) %>%
+            select(!!theClasses[i + 1] := label, id = has_broader, new_id = id)
 
-        toOut <- toOut %>%
-          filter(!label %in% temp$external_label)
+          temp <- temp %>%
+            left_join(child, by = "id")
 
-        temp <- temp %>%
-          group_by(id, label, class, description, has_broader, match) %>%
-          summarise(external_label = paste0(external_label, collapse = " | ")) %>%
-          ungroup() %>%
-          pivot_wider(id_cols = c(id, label, class, description, has_broader), names_from = match, values_from = external_label)
+          if(i == 1){
+            flatOnto <- temp %>%
+              mutate(!!paste0(theClasses[i], "_id") := id,
+                     !!paste0(theClasses[i + 1], "_id") := new_id) %>%
+              select(theClasses[i], paste0(theClasses[i], "_id"), theClasses[i + 1], paste0(theClasses[i + 1], "_id"), new_id)
+          } else {
+            flatOnto <- flatOnto %>%
+              mutate(id = new_id) %>%
+              select(-new_id) %>%
+              left_join(temp %>% select(id, theClasses[i + 1], new_id), by = "id") %>%
+              mutate(!!paste0(theClasses[i + 1], "_id") := new_id)
+          }
+        }
 
-        toOut <- toOut %>%
-          bind_rows(temp) %>%
-          arrange(id)
+        flatOnto <- flatOnto %>%
+          select(-id, -new_id) %>%
+          select(contains(colnames(table)))
+
+        targetClass <- tail(colnames(table), 1)
+        parentClass <- tail(colnames(table), 2)[1]
+
+        toOut <- table %>%
+          left_join(flatOnto, by = colnames(table)) %>%
+          distinct() %>%
+          select(label = {{ targetClass }}, id = paste0(targetClass, "_id")) %>%
+          mutate(class = targetClass)
+
+        parentOut <- table %>%
+          left_join(flatOnto %>% distinct(!!sym(parentClass), !!sym(paste0(parentClass, "_id"))), by = all_of(parentClass)) %>%
+          select(label = all_of(targetClass), has_broader = paste0(parentClass, "_id"))
+
+        toOut <- ontology@concepts$harmonised %>%
+          filter(id %in% toOut$id) %>%
+          full_join(toOut, by = c("id", "label", "class"))
+
+        if(all(is.na(toOut$has_broader))){
+          toOut <- toOut %>%
+            select(-has_broader) %>%
+            left_join(parentOut, by = "label") %>%
+            select(id, label, class, has_broader, everything())
+        }
+
+      } else {
+
+        assertNames(x = names(table), subset.of = c("id", "class", "label", "description", "has_broader", "has_source"))
+
+        if(!regex){
+
+          toOut <- table %>%
+            left_join(ontology@concepts$harmonised, by = colnames(table)) %>%
+            mutate(external = label,
+                   match = "exact",
+                   has_source = "1") %>%
+            select(external, match, label, class, id, has_broader, description, has_source)
+
+          if("label" %in% names(table)){
+
+            subsNames <- colnames(table)[colnames(table) %in% names(ontology@concepts$external)]
+
+            extOut <- table %>%
+              left_join(ontology@concepts$external, by = subsNames) %>%
+              select(extid = id, extLabel = label, has_source) %>%
+              filter(!is.na(extid))
+
+            if(dim(extOut)[1] != 0){
+
+              extOut <- ontology@concepts$harmonised %>%
+                pivot_longer(cols = c(has_close_match, has_broader_match, has_narrower_match, has_exact_match), names_to = "match", values_to = "extid") %>%
+                separate_rows(extid, sep = " \\| ") %>%
+                separate(col = extid, into = c("extid", "certainty"), sep = "[.]") %>%
+                filter(extid %in% na.omit(extOut$extid)) %>%
+                left_join(extOut, ., by = "extid") %>%
+                mutate(match = str_replace_all(match, "has_", ""),
+                       match = str_replace_all(match, "_match", "")) %>%
+                select(external = extLabel, match, label, class, id, has_broader, description, has_source)
+
+            } else {
+              extOut <- extOut %>%
+                select(external = extLabel, has_source)
+            }
+
+            # rename for join
+            table <-  table %>%
+              select(external = label, everything()) %>%
+              distinct()
+
+            toOut <- toOut %>%
+              bind_rows(extOut) %>%
+              arrange(match) %>%
+              left_join(table, ., by = colnames(table))
+          }
+
+        } else {
+
+          toOut <- ontology@concepts$harmonised
+
+          for(j in seq_along(table)){
+            toOut <- toOut %>%
+              filter(str_detect(string = toOut[[colnames(table)[j]]],
+                                pattern = paste0(c(table %>% pull(j)), collapse = "|")))
+          }
+
+          regexpr <- paste0(paste0(names(table), "~", table), collapse = ",")
+          toOut <- toOut %>%
+            mutate(external = regexpr,
+                   match = "regex",
+                   has_source = "1") %>%
+            select(external, match, label, class, id, has_broader, description, has_source)
+        }
+
+        if(mappings){
+
+          toOut <- toOut %>%
+            distinct(external, class, id, has_broader) %>%
+            left_join(ontology@concepts$harmonised, by = c("class", "id", "has_broader"))
+
+          toOut <- toOut %>%
+            pivot_longer(cols = c(has_close_match, has_broader_match, has_narrower_match, has_exact_match), names_to = "match", values_to = "extid") %>%
+            separate_rows(extid, sep = " \\| ") %>%
+            separate(col = extid, into = c("extid", "certainty"), sep = "[.]") %>%
+            left_join(ontology@concepts$external %>% select(extid = id, external_label = label), by = "extid") %>%
+            group_by(label, class, id, description, has_broader, match) %>%
+            summarise(external_label = paste0(unique(external_label), collapse = " | ")) %>%
+            ungroup() %>%
+            mutate(external_label = na_if(external_label, "NA")) %>%
+            pivot_wider(id_cols = c(label, class, id, description, has_broader), names_from = match, values_from = external_label)
+
+          if(!regex){
+            toOut <- toOut %>%
+              rename(external = label) %>%
+              left_join(table, ., by = colnames(table)) %>%
+              rename(label = external)
+          }
+
+        }
+
       }
 
-    } else {
-      toOut <- toOut %>%
-        select(id, label, class, description, has_broader) %>%
-        arrange(id)
     }
-
-    for(i in seq_along(attrib)){
-
-      toOut <- toOut %>%
-        filter(str_detect(toOut[[names(attrib)[i]]], paste0(eval_tidy(attrib[[i]]), collapse = "|")))
-
-    }
-
   }
 
-  # if(missing){
-
-    # out <- toOut %>%
-      # filter(is.na(id))
-
-  # } else {
-
-    if(tree){
-
-      topID <- toOut %>%
-        pull(id) %>%
-        unique()
-
-      out <- make_tree(theConcepts$harmonised, topID)
-
-    } else {
-
-      if(na.rm){
-        out <- toOut %>%
-          filter(!is.na(id))
-      } else {
-        out <- toOut
-      }
-    }
-
-  out <- out %>%
-    select(label, class, id, has_broader, description, everything())
-
-  # }
-
-  return(out)
+  return(toOut)
 
 }