+++ inst/doc/map_new_concepts.R
@@ -22,14 +22,14 @@ crops <- new_source(name = "externalDataset",
                     ontology = crops)
 
 # new concepts that occur in the external dataset, which should be harmonised with the ontology
-newConcepts <- tibble(label = c("Wheat", "NUTS", "Avocado"))
+externalConcepts <- tibble(label = c("Wheat", "NUTS", "Avocado"))
 
 ## ----get_concepts missing-----------------------------------------------------
-missingConcepts <- get_concept(x = newConcepts, ontology = crops)
+missingConcepts <- get_concept(table = externalConcepts, ontology = crops)
 kable(missingConcepts)
 
 ## ----set_concept--------------------------------------------------------------
-broaderConcepts <- get_concept(x = tibble(label = c("Wheat", "Tropical and subtropical Fruit")), 
+broaderConcepts <- get_concept(table = tibble(label = c("Wheat", "Tropical and subtropical Fruit")), 
                                ontology = crops)
 
 crops <- new_concept(new = c("wheat", "avocado"),
@@ -38,10 +38,10 @@ crops <- new_concept(new = c("wheat", "avocado"),
                      ontology = crops)
 
 ## ----set_mapping--------------------------------------------------------------
-toMap <- get_concept(x = tibble(label = c("wheat", "NUTS", "avocado")),
+toMap <- get_concept(table = tibble(label = c("wheat", "NUTS", "avocado")),
                      ontology = crops)
 
-crops <- new_mapping(new = newConcepts %>% pull(label),
+crops <- new_mapping(new = externalConcepts %>% pull(label),
                      target = toMap,
                      match = c("close", "close", "close"),
                      source = "externalDataset",
@@ -49,7 +49,7 @@ crops <- new_mapping(new = newConcepts %>% pull(label),
                      ontology = crops)
 
 ## -----------------------------------------------------------------------------
-broaderConcepts <- get_concept(x = tibble(label = c("wheat", "wheat")),
+broaderConcepts <- get_concept(table = tibble(label = c("wheat", "wheat")),
                                ontology = crops)
 
 # for (some of) these concepts we do not know the class ...
@@ -58,8 +58,11 @@ crops <- new_concept(new = c("wheat1", "wheat2"),
                      class = NA_character_,
                      ontology = crops)
 
-get_concept(x = tibble(label = "Wheat"), tree = TRUE, ontology = crops)[,1:6] %>% 
+get_concept(table = tibble(label = "Wheat"), ontology = crops) %>%
+  make_tree(ontology = crops) %>% 
+  select(1:5) %>% 
   kable()
+
 # ... ok, then let's specify that class and re-run new_concept
 crops <- new_class(new = "cultivar", target = "crop", 
                    description = "type of plant that people have bred for desired traits", 
@@ -71,12 +74,22 @@ crops <- new_concept(new = c("wheat1", "wheat2"),
                      ontology = crops)
 
 ## ----new ontology-------------------------------------------------------------
-get_concept(x = tibble(label = "Wheat"), tree = TRUE, ontology = crops)[,1:6] %>% 
+get_concept(table = tibble(label = "Wheat"), ontology = crops) %>%
+  make_tree(ontology = crops) %>% 
+  select(1:6) %>% 
   kable()
 
-get_concept(x = tibble(label = "NUTS"), tree = TRUE, ontology = crops)[,1:6] %>% 
+get_concept(table = tibble(label = "NUTS"), ontology = crops) %>% 
+  make_tree(ontology = crops) %>% 
+  select(1:6) %>% 
   kable()
 
-get_concept(x = tibble(label = "FRUIT"), tree = TRUE, ontology = crops)[,1:6] %>% 
+get_concept(table = tibble(label = "FRUIT"), ontology = crops) %>% 
+  make_tree(ontology = crops) %>% 
+  select(1:6) %>% 
+  kable()
+
+# and finally a list of all external concepts
+get_concept(external = TRUE, ontology = crops) %>% 
   kable()